# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LA RC Cafe API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all shop products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, description, price, category, featured, inStock, stock]
      properties:
        id:          { type: integer }
        name:        { type: string }
        description: { type: string }
        price:       { type: integer }
        category:    { type: string }
        featured:    { type: boolean }
        inStock:     { type: boolean }
        stock:       { type: integer }
        imageUrl:    { type: ["string", "null"] }
    ProductInput:
      type: object
      required: [name, description, price, category]
      properties:
        name:        { type: string, minLength: 1 }
        description: { type: string }
        price:       { type: integer }
        category:    { type: string, minLength: 1 }
        featured:    { type: boolean }
        inStock:     { type: boolean }
        stock:       { type: integer }
        imageUrl:    { type: string }
    ProductUpdate:
      type: object
      properties:
        name:        { type: string }
        description: { type: string }
        price:       { type: integer }
        category:    { type: string }
        featured:    { type: boolean }
        inStock:     { type: boolean }
        stock:       { type: integer }
        imageUrl:    { type: string }
...